% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_trial}
\alias{in_trial}
\title{Identify which population rows are included in a trial subset}
\usage{
in_trial(population, trial)
}
\arguments{
\item{population}{A data frame or a matrix
representing the covariates of population}

\item{trial}{A data frame or a matrix
representing the covariates of trial dataset, assumed to be a subset of \code{population}
by rows.}
}
\value{
A logical vector of length \code{nrow(population)}. The \code{i}-th
element is \code{TRUE} if the \code{i}-th row of \code{population} appears
in \code{trial}, and \code{FALSE} otherwise.
}
\description{
Returns a logical vector indicating, for each row in \code{population},
whether that row is included in \code{trial}, where \code{trial} is assumed
to be a (row) subset of \code{population}.
}
\examples{
pop <- data.frame(id = 1:5, x = c(0.1, 0.2, 0.3, 0.4, 0.5))
tri <- pop[c(2, 4), ]
in_trial(pop, tri)

}
